zombie-shooter/backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
channels==4.3.1
psycopg2-binary==2.9.10